requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0

google-generativeai==0.5.4

beautifulsoup4==4.13.4
requests==2.32.5

python-dotenv==1.1.1
pydantic-settings==2.10.1

google-cloud-firestore